Force PNG output
The tool is more composable and unix-y if the output is always png, for (at least) two reasons: 1. The output is predictable, and the tool can appear in pipelines that can not automatically tell a jpeg from a png, 2. png is lossless. This means that the tool does one thing: renders a snapshot of the page. If the user wants lossy compression (that's created for photographic content), she can easily add that later.